+++ extension/install_paddle_ocr/meta.json
@@ -1,0 +1,7 @@
+{
+    "extension_name": "Install PaddleOCR Module for Windrecorder",
+    "developer_name": ["ASC8384", "antonoko"],
+    "developer_url": ["https://github.com/ASC8384", "https://github.com/Antonoko"],
+    "version": "0.0.1",
+    "description_markdown": "add Paddle OCR as optional OCR engine for Windrecorder"
+}
