--- extension/paddle_ocr/meta.json
@@ -1,7 +1,0 @@
-{
-    "extension_name": "Install PaddleOCR Module for Windrecorder",
-    "developer_name": "ASC8384",
-    "developer_url": "https://github.com/ASC8384",
-    "version": "0.0.1",
-    "description_markdown": ""
-}
